Move putEventStringInData to subclass Event
