+++ app/src/main/java/com/github/houseorganizer/houseorganizer/util/Util.java
@@ -11,29 +11,9 @@ import android.widget.Toast;
 
 import com.github.houseorganizer.houseorganizer.R;
 
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.TemporalAccessor;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 
 public class Util {
-
-
-    public static boolean putEventStringsInData(Map<String, String> event, Map<String, Object> data) {
-        data.put("title", event.get("title"));
-        data.put("description", event.get("desc"));
-        try {
-            TemporalAccessor start = DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm").parse(event.get("date"));
-            data.put("start", LocalDateTime.from(start).toEpochSecond(ZoneOffset.UTC));
-            data.put("duration", Integer.valueOf(Objects.requireNonNull(event.get("duration"))));
-        } catch(Exception e) {
-            return true;
-        }
-        return false;
-    }
 
     //<-------------------¦ Preferences ¦--------------------------------------------------->
